+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Upgrade `@folio/stripes-acq-components` to `v7.0.2`. Refs UIORGS-466.
 * Memoize initial values for the integration form to prevent data from becoming stale. Refs UIORGS-472.
 * Include global `mod-settings` permissions in base permission sets. Refs UIORGS-476.
+* Replace the `moment` library with `dayjs` (Part 1). Refs UIORGS-463.
 
 ## [6.0.1](https://github.com/folio-org/ui-organizations/tree/v6.0.1) (2025-05-05)
 [Full Changelog](https://github.com/folio-org/ui-organizations/compare/v6.0.0...v6.0.1)